Pulled out openfin chart epic to the openfin folder
