+++ src/client/src/services/openFin/epics.ts
@@ -1,8 +1,10 @@
 import { combineEpics } from 'redux-observable'
 import { connectAnalyticsServiceToOpenFinEpic } from './epics/analyticsServiceEpic'
 import { connectBlotterServiceToOpenFinEpic } from './epics/blotterServiceEpic'
+import { displayCurrencyChartEpic } from './epics/displayCurrencyChartEpic'
 
 export const openfinServiceEpics = combineEpics(
   connectAnalyticsServiceToOpenFinEpic,
-  connectBlotterServiceToOpenFinEpic
+  connectBlotterServiceToOpenFinEpic,
+  displayCurrencyChartEpic
 )
