+++ src/client/src/ui/spotTile/epics.ts
@@ -1,6 +1,5 @@
 import { Action } from 'redux'
 import { combineEpics, ofType } from 'redux-observable'
-import { from as observableFrom } from 'rxjs'
 import { delay, map, mergeMap, tap } from 'rxjs/operators'
 import { ApplicationEpic } from '../../ApplicationEpic'
 import { ExecuteTradeResponse } from '../../types'
@@ -25,18 +24,6 @@ const executeTradeEpic: ApplicationEpic = (action$, state$, { executionService }
     )
   )
 
-type DisplayChartAction = ReturnType<typeof SpotTileActions.displayCurrencyChart>
-type ChartOpenedAction = ReturnType<typeof SpotTileActions.currencyChartOpened>
-
-export const displayCurrencyChart: ApplicationEpic = (action$, state$, { openFin }) =>
-  action$.pipe(
-    ofType<Action, DisplayChartAction>(ACTION_TYPES.DISPLAY_CURRENCY_CHART),
-    mergeMap<DisplayChartAction, string>((action: DisplayChartAction) =>
-      observableFrom<string>(openFin.displayCurrencyChart(action.payload))
-    ),
-    map<string, ChartOpenedAction>(symbol => SpotTileActions.currencyChartOpened(symbol))
-  )
-
 type DismissNotificationAction = ReturnType<typeof SpotTileActions.dismissNotification>
 
 export const onTradeExecuted: ApplicationEpic = (action$, state$, { openFin }) =>
@@ -52,4 +39,4 @@ export const onTradeExecuted: ApplicationEpic = (action$, state$, { openFin }) =
     map<string, DismissNotificationAction>(SpotTileActions.dismissNotification)
   )
 
-export const spotTileEpicsCreator = combineEpics(executeTradeEpic, displayCurrencyChart, onTradeExecuted)
+export const spotTileEpicsCreator = combineEpics(executeTradeEpic, onTradeExecuted)
