+++ src/client/src/ui/spotTile/index.ts
@@ -6,3 +6,4 @@ export { default as NotionalInput } from './notional/NotionalInput'
 export { default as SpotTileControls } from './spotTileControls/SpotTileControls'
 export { default as spotTileDataReducer } from './reducer'
 export { spotTileEpicsCreator } from './epics'
+export { ACTION_TYPES, SpotTileActions } from './actions'
